api/workers/requirements.txt
# EEG Signal Processing
mne==1.8.0
numpy>=1.26.0,<2.0.0
scipy>=1.13.0,<1.14.0  # MNE 1.8.0 requires sph_harm which was removed in scipy 1.14.0
pandas>=2.1.0

# Artifact Detection & Complexity
scikit-learn>=1.4.0
antropy==0.1.6

# Visualization
matplotlib>=3.8.0
Pillow>=10.0.0

# PDF Generation
reportlab==4.0.9

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Supabase Integration
supabase==2.10.0
python-dotenv==1.0.0

# Web Server
Flask==3.0.0
gunicorn==21.2.0

# Database (for local/Docker mode)
psycopg2-binary==2.9.9
